added GenerateDescription
