+++ frontend/src/components/pages/MainComponent/MyModal/MyModal.jsx
@@ -1,5 +1,4 @@
 import { useRef } from "react";
-import api from "./../../../utils/api";
 import MyModalFrom from "./MyModalForm";
 
 export default function MyModal({
@@ -10,24 +9,6 @@ export default function MyModal({
 }) {
     const buttonRef = useRef();
 
-    async function doRequestToGigachat() {
-        try {
-            const title = form.getFieldValue("title");
-
-            const response = await api.get("/gigachat/get-description", {
-                params: { title: title },
-            });
-
-            form.setFieldsValue({
-                description: response.data.description,
-            });
-
-            buttonRef.current?.blur();
-        } catch (error) {
-            console.error("Ошибка в запросе с GigaChat:", error);
-        }
-    }
-
     return (
         <MyModalFrom
             form={form}
@@ -35,7 +16,6 @@ export default function MyModal({
             isModalVisible={isModalVisible}
             setIsModalVisible={setIsModalVisible}
             buttonRef={buttonRef}
-            doRequestToGigachat={doRequestToGigachat}
         ></MyModalFrom>
     );
 }
